requirements.txt
cryptography
python-dotenv 
flask 
flask_jwt_extended 
flask_smorest 